// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Categoria: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly ItemVenda: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly vendaId: CodecTypes['pg/int4@1']['output'];
      readonly produtoId: CodecTypes['pg/int4@1']['output'];
      readonly quantidade: CodecTypes['pg/int4@1']['output'];
      readonly precoUnitario: CodecTypes['pg/numeric@1']['output'];
      readonly subtotal: CodecTypes['pg/numeric@1']['output'];
    };
    readonly MovimentoStock: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly produtoId: CodecTypes['pg/int4@1']['output'];
      readonly utilizadorId: CodecTypes['pg/int4@1']['output'];
      readonly tipo: 'ENTRADA' | 'SAIDA' | 'AJUSTE';
      readonly quantidade: CodecTypes['pg/int4@1']['output'];
      readonly stockAnterior: CodecTypes['pg/int4@1']['output'];
      readonly stockPosterior: CodecTypes['pg/int4@1']['output'];
      readonly motivo: CodecTypes['pg/text@1']['output'] | null;
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Produto: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly codigo: CodecTypes['pg/text@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly categoriaId: CodecTypes['pg/int4@1']['output'];
      readonly precoCompra: CodecTypes['pg/numeric@1']['output'];
      readonly precoVenda: CodecTypes['pg/numeric@1']['output'];
      readonly stockActual: CodecTypes['pg/int4@1']['output'];
      readonly stockMinimo: CodecTypes['pg/int4@1']['output'];
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Utilizador: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly palavraPasse: CodecTypes['pg/text@1']['output'];
      readonly perfil: 'ADMINISTRADOR' | 'ATENDENTE';
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Venda: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly utilizadorId: CodecTypes['pg/int4@1']['output'];
      readonly total: CodecTypes['pg/numeric@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Categoria: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly descricao: CodecTypes['pg/text@1']['input'] | null;
      readonly activo: CodecTypes['pg/bool@1']['input'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly ItemVenda: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly vendaId: CodecTypes['pg/int4@1']['input'];
      readonly produtoId: CodecTypes['pg/int4@1']['input'];
      readonly quantidade: CodecTypes['pg/int4@1']['input'];
      readonly precoUnitario: CodecTypes['pg/numeric@1']['input'];
      readonly subtotal: CodecTypes['pg/numeric@1']['input'];
    };
    readonly MovimentoStock: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly produtoId: CodecTypes['pg/int4@1']['input'];
      readonly utilizadorId: CodecTypes['pg/int4@1']['input'];
      readonly tipo: 'ENTRADA' | 'SAIDA' | 'AJUSTE';
      readonly quantidade: CodecTypes['pg/int4@1']['input'];
      readonly stockAnterior: CodecTypes['pg/int4@1']['input'];
      readonly stockPosterior: CodecTypes['pg/int4@1']['input'];
      readonly motivo: CodecTypes['pg/text@1']['input'] | null;
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Produto: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly codigo: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly descricao: CodecTypes['pg/text@1']['input'] | null;
      readonly categoriaId: CodecTypes['pg/int4@1']['input'];
      readonly precoCompra: CodecTypes['pg/numeric@1']['input'];
      readonly precoVenda: CodecTypes['pg/numeric@1']['input'];
      readonly stockActual: CodecTypes['pg/int4@1']['input'];
      readonly stockMinimo: CodecTypes['pg/int4@1']['input'];
      readonly activo: CodecTypes['pg/bool@1']['input'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Utilizador: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly palavraPasse: CodecTypes['pg/text@1']['input'];
      readonly perfil: 'ADMINISTRADOR' | 'ATENDENTE';
      readonly activo: CodecTypes['pg/bool@1']['input'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Venda: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly utilizadorId: CodecTypes['pg/int4@1']['input'];
      readonly total: CodecTypes['pg/numeric@1']['input'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly categoria: {
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
    };
    readonly itemVenda: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly precoUnitario: CodecTypes['pg/numeric@1']['output'];
      readonly produtoId: CodecTypes['pg/int4@1']['output'];
      readonly quantidade: CodecTypes['pg/int4@1']['output'];
      readonly subtotal: CodecTypes['pg/numeric@1']['output'];
      readonly vendaId: CodecTypes['pg/int4@1']['output'];
    };
    readonly movimentoStock: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly motivo: CodecTypes['pg/text@1']['output'] | null;
      readonly produtoId: CodecTypes['pg/int4@1']['output'];
      readonly quantidade: CodecTypes['pg/int4@1']['output'];
      readonly stockAnterior: CodecTypes['pg/int4@1']['output'];
      readonly stockPosterior: CodecTypes['pg/int4@1']['output'];
      readonly tipo: 'ENTRADA' | 'SAIDA' | 'AJUSTE';
      readonly utilizadorId: CodecTypes['pg/int4@1']['output'];
    };
    readonly produto: {
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly categoriaId: CodecTypes['pg/int4@1']['output'];
      readonly codigo: CodecTypes['pg/text@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly precoCompra: CodecTypes['pg/numeric@1']['output'];
      readonly precoVenda: CodecTypes['pg/numeric@1']['output'];
      readonly stockActual: CodecTypes['pg/int4@1']['output'];
      readonly stockMinimo: CodecTypes['pg/int4@1']['output'];
    };
    readonly utilizador: {
      readonly activo: CodecTypes['pg/bool@1']['output'];
      readonly actualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly palavraPasse: CodecTypes['pg/text@1']['output'];
      readonly perfil: 'ADMINISTRADOR' | 'ATENDENTE';
    };
    readonly venda: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly total: CodecTypes['pg/numeric@1']['output'];
      readonly utilizadorId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;